+++ spec/lib/attr_digest_spec.rb
@@ -1,5 +1,151 @@
 require 'spec_helper'
 
+describe ModelWithUsername do
+  it 'responds to :password' do
+    respond_to(:password)
+  end
+
+  it 'responds to :password=' do
+    respond_to(:password=)
+  end
+
+  it 'responds to :authenticate_password' do
+    respond_to(:authenticate_password)
+  end
+
+  it 'raises NoDigestException when digest is non-existent' do
+    model = FactoryGirl.create(:model_with_username, username: 'Bam Bam')
+    expect(lambda do
+      model.authenticate_password('abc')
+    end).to raise_error(AttrDigest::NoDigestException)
+  end
+
+  it 'returns true if authenticate_password is passed an empty string' do
+    model = FactoryGirl.create(:model_with_username, username: 'Fred', password: '')
+    expect(model.authenticate_password('')).to be(true)
+  end
+
+  it 'returns true if authenticate_password is passed a nil' do
+    model = FactoryGirl.create(:model_with_username, username: 'Barney', password: '')
+    expect(model.authenticate_password(nil)).to be(true)
+  end
+end
+
+describe ModelWithFormatOption do
+  it 'fails validation if password does not meet format conditions' do
+    model = FactoryGirl.create(:model_with_format_option)
+    model.password = "p455w0rd"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("only allows letters")
+  end
+
+  it 'passes validation if password meets format conditions' do
+    model = FactoryGirl.create(:model_with_format_option, username: 'Rocky Snr')
+    model.password = "password"
+    expect(model.valid?).to be(true)
+  end
+end
+
+describe ModelWithMinimumLengthOption do
+  it 'fails validation if password does not meet minimum length condition' do
+    model = FactoryGirl.create(:model_with_minimum_length_option)
+    model.password = "pass"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("is too short (minimum is 5 characters)")
+  end
+
+  it 'passes validation if password meets minimum length condition' do
+    model = FactoryGirl.create(:model_with_minimum_length_option, username: 'Pebbles Jnr')
+    model.password = "password"
+    expect(model.valid?).to be(true)
+  end
+end
+
+describe ModelWithMaximumLengthOption do
+  it 'fails validation if password does not meet maximum length condition' do
+    model = FactoryGirl.create(:model_with_maximum_length_option)
+    model.password = "extra-long-password"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("is too long (maximum is 10 characters)")
+  end
+
+  it 'passes validation if password meets maximum length condition' do
+    model = FactoryGirl.create(:model_with_maximum_length_option, username: 'The late Mr Slate')
+    model.password = "password"
+    expect(model.valid?).to be(true)
+  end
+end
+
+describe ModelWithLengthRangeOption do
+  it 'fails validation if password does not meet ranged length condition (too long)' do
+    model = FactoryGirl.create(:model_with_length_range_option)
+    model.password = "extra-long-password"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("is too long (maximum is 10 characters)")
+  end
+
+  it 'fails validation if password does not meet ranged length condition (too short)' do
+    model = FactoryGirl.create(:model_with_length_range_option, username: 'Arnold Jnr')
+    model.password = "pass"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("is too short (minimum is 5 characters)")
+  end
+
+  it 'passes validation if password meets ranged length condition' do
+    model = FactoryGirl.create(:model_with_length_range_option, username: 'Arnold Snr')
+    model.password = "password"
+    expect(model.valid?).to be(true)
+  end
+end
+
+describe ModelWithExactLengthOption do
+  it 'fails validation if password does not meet exact length condition' do
+    model = FactoryGirl.create(:model_with_exact_length_option)
+    model.password = "short"
+    expect(model.valid?).to be(false)
+    expect(model.errors[:password]).to include("is the wrong length (should be 8 characters)")
+  end
+
+  it 'passes validation if password meets exact length condition' do
+    model = FactoryGirl.create(:model_with_exact_length_option, username: 'The late Grand Poobah')
+    model.password = "password"
+    expect(model.valid?).to be(true)
+  end
+end
+
+describe ModelWithTimeAndMemoryCostOptions do
+  it 'creates a digest with alternative time and memory costs' do
+    model = FactoryGirl.create(:model_with_time_and_memory_cost_options, username: 'Wilma', password: 'abc')
+    expect(model.authenticate_password('abc')).to be(true)
+    expect(model.password_digest).to match(/m=4096,t=3/)
+  end
+end
+
+describe ModelWithInvalidMemoryCostOption do
+  it 'raises InvalidMemoryCost' do
+    model = FactoryGirl.create(:model_with_invalid_memory_cost_option, username: 'Betty')
+    expect(lambda do
+      model.password = "abc"
+    end).to raise_error(AttrDigest::InvalidMemoryCost)
+  end
+end
+
+describe ModelWithInvalidTimeCostOption do
+  it 'raises InvalidTimeCost' do
+    model = FactoryGirl.create(:model_with_invalid_time_cost_option, username: 'Dino')
+    expect(lambda do
+      model.password = "abc"
+    end).to raise_error(AttrDigest::InvalidTimeCost)
+  end
+end
+
+describe ModelWithAttrDigestSecretOption do
+  it 'permits setting the secret option' do
+    model = FactoryGirl.create(:model_with_attr_digest_secret_option)
+    expect(model.authenticate_password('password')).to be(true)
+  end
+end
+
 describe ModelWithAttrDigest do
   it 'responds to :security_answer' do
     respond_to(:security_answer)
@@ -112,7 +258,6 @@ describe ModelWithAttrDigest do
       expect(model.authenticate_security_answer('some other answer')).to be(false)
     end
   end
-
 end
 
 describe ModelWithAttrDigestAndValidationsOption do